requirements.txt
fastapi[standard]
sqlalchemy
jinja2>=3.1.0
python-multipart
passlib[bcrypt]
fastapi-sessions
pyjwt
pydantic_settings


